Make inherited Runner logic reusable for specs as well
